Rename "shared" components to "common" components
