+++ app/javascript/components/Board/SearchFilter.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import I18n from 'i18n-js';
 
-import SidebarBox from '../shared/SidebarBox';
+import SidebarBox from '../common/SidebarBox';
 
 interface Props {
   searchQuery: string;
