+++ app/javascript/components/Post/ActionBox.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import I18n from 'i18n-js';
 
-import Button from '../shared/Button';
-import { SmallMutedText } from '../shared/CustomTexts';
-import SidebarBox from '../shared/SidebarBox';
+import Button from '../common/Button';
+import { SmallMutedText } from '../common/CustomTexts';
+import SidebarBox from '../common/SidebarBox';
 
 interface Props {
   followed: boolean;
