+++ app/javascript/components/Post/PostP.tsx
@@ -12,11 +12,11 @@ import ActionBox from './ActionBox';
 import LikeButton from '../../containers/LikeButton';
 import PostBoardSelect from './PostBoardSelect';
 import PostStatusSelect from './PostStatusSelect';
-import PostBoardLabel from '../shared/PostBoardLabel';
-import PostStatusLabel from '../shared/PostStatusLabel';
+import PostBoardLabel from '../common/PostBoardLabel';
+import PostStatusLabel from '../common/PostStatusLabel';
 import Comments from '../../containers/Comments';
-import { MutedText } from '../shared/CustomTexts';
-import Sidebar from '../shared/Sidebar';
+import { MutedText } from '../common/CustomTexts';
+import Sidebar from '../common/Sidebar';
 
 import { LikesState } from '../../reducers/likesReducer';
 import { CommentsState } from '../../reducers/commentsReducer';
