+++ app/javascript/components/common/CustomTexts.tsx
