+++ app/javascript/components/common/DragZone.tsx
