+++ app/javascript/components/common/PostBoardLabel.tsx
