+++ app/javascript/components/common/Separator.tsx
