+++ app/javascript/components/common/SidebarBox.tsx
